Rename OSNotification class
Motivation: avoid conflict with native Notification type
